+++ handlers/kinetic_request_ce_submission_db_insert_v1/handler/init.rb
@@ -39,7 +39,7 @@ class KineticRequestCeSubmissionDbInsertV1
       :columnName   => 128,
       :coreState    => 10,
       :createdBy    => 255,
-      :fieldKey     => 36,
+      :fieldKey     => 255,
       :formField    => 4000,
       :formName     => 255,
       :formSlug     => 255,
@@ -139,7 +139,7 @@ class KineticRequestCeSubmissionDbInsertV1
       @table_temp_prefix.prepend("#")
       #@db.transaction_isolation_level = :serializable
       #@db.transaction_isolation_level = :repeatable
-      #@db.transaction_isolation_level = :committed
+      #@db.transaction_isolation_level = :committed      
     elsif @info_values["jdbc_database_id"].downcase == "oracle"
       Sequel.database_timezone = :utc
       #Sequel.application_timezone = :utc
@@ -243,7 +243,7 @@ class KineticRequestCeSubmissionDbInsertV1
         :api_username => api_username, 
         :api_password => api_password
       })
-      # Update the kapp table with new kapp feilds
+      # Update the kapp table with new kapp fields
       update_kapp_table_columns({
         :kapp_slug => kapp_slug
       })
@@ -353,6 +353,12 @@ class KineticRequestCeSubmissionDbInsertV1
             .map {|k,v| 
               {k.to_sym => v}
             }.reduce Hash.new, :merge
+
+          #HOTFIX - Update fieldKey size
+          fieldKeySize = check_field_size(form_table_name, 'fieldKey')
+          if fieldKeySize == 8
+            alter_column_type_size(form_table_name, 'fieldKey', 'varchar', @db_column_size_limits[:fieldKey])
+          end
 
           @db[form_table_name.to_sym].call(
             :insert,
@@ -532,7 +538,7 @@ class KineticRequestCeSubmissionDbInsertV1
             .map {|k,v| 
               {k.to_sym => v}
             }.reduce Hash.new, :merge
-
+          
           # if the record does not exist in the database, insert it.
           if datastore != "yes" then
             if @info_values['first_bulk_load'] || @db[kapp_table_name.to_sym].select(:c_id).where(:c_id => submission["id"]).count == 0 then
@@ -1292,4 +1298,38 @@ class KineticRequestCeSubmissionDbInsertV1
   end
   # This is a ruby constant that is used by the escape method
   ESCAPE_CHARACTERS = {'&'=>'&amp;', '>'=>'&gt;', '<'=>'&lt;', '"' => '&quot;'}
+end
+
+##########################################################################################################
+#
+# check_field_size
+#
+# Returns size of field, created to resolve fieldKey size change in v6 upgrade
+#
+##########################################################################################################
+
+def check_field_size(tableName, fieldName)
+  size = nil
+  @db.schema(tableName.to_sym).each {|label,columnDetails| 
+    if label.to_s == fieldName
+      #Return size
+      size = columnDetails[:max_chars]
+      return size
+    end
+  }
+  return nil
+end
+
+##########################################################################################################
+#
+# alter_column_type_size
+#
+# Alters column type and or size
+#
+##########################################################################################################
+
+def alter_column_type_size(tableName, fieldName, dataType, fieldSize)
+  @db.alter_table(tableName.to_sym) do
+    set_column_type(fieldName.to_sym, dataType.to_sym, size: fieldSize)
+  end
 end