+++ handlers/kinetic_request_ce_submission_db_bulk_upsert_v1/CHANGELOG.md
@@ -1,2 +1,4 @@
+Update - v6 fieldKey hotfix (2024-12-10)
+ * Hotfix added to increase fieldKey column size to accomodate for GUIDs in v6 being the default entry of the field
 Kinetic Reporting DB Upsert V1 (2021-06-12)
  * Initial version.  See README for details.